Add files via upload
So far, South Africa has recorded the highest number of COVID-19 cases in Africa. A large percentage of the population also have pre-existing comorbidities such as HIV, TB and diabetes which made the handling of these COVID-19 cases to be more challenging to manage. Recent studies have shown that Artificial Intelligence can help health practitioners to make informed decisions when treating COVID-19 patients. This paper reports the application of three machine learning (ML) algorithms: Multi-layer Perceptron (MLP), Extreme Boosted Trees (XGBoost) and Support Vector Machines (SVM) for predicting the outcome of intensive care patients of COVID-19 with comorbidities in a South African hospital. We also report our findings on the performance of the three ML models  when cross-validation (CV), and principal component analysis (PCA) for dimensionality reduction were applied on the COVID-19 data set. The results in terms of F1 score and AUC score show that MLP had the best overall performance when only CV was applied without PCA (F1=0.98; AUC = 0.98). We also found that the MLP generally performed better when CV was applied, and had a good performance without CV. MLP is also affected by PCA and CV. SVM generally performed better with CV and is affected by PCA. XGBoost had the best performance when none of cross-validation and PCA was used (F1=0.96; AUC = 0.97). It is not affected by CV and performed worse with PCA. Generally, the performance of both SVM and MLP can be enhanced through cross-validation, and they also perform very well without PCA. Apart from demonstrating how ML can be applied for the management of COVID-19 cases, the findings of the study offer a guide for researchers on treatments that could be applied to the selected ML models when used for classification tasks that relate to prediction of patients’ outcomes.