Predicting COVID-19 Breakthrough Infection
So far, South Africa has recorded the highest number of COVID-19 cases in Africa. A large percentage of the population also have pre-existing comorbidities such as HIV, TB, and diabetes. The existence of comorbidities in patients with COVID-19 complicates prevention strategies and making it more difficult to manage. With the emergence of new variants of the COVID-19 virus, vaccinated persons with pre-existing comorbidities are still prone to COVID-19 re-infection and breakthrough infections. Although vaccines are not expected to provide 100% protection, they do offer different levels of immune protection. For example, emerging data suggests that  some COVID-19 vaccines have a higher rate of breakthrough infections compared to other vaccines. This research focuses on predicting breakthrough infection using machine learning so as to determine the top factors or variables that determine breakthrough infection. The study seeks to facilitate guidance on clinical decision-making for patients with a history of prior infection of COVID-19 using the methods of Explainable Artificial Intelligence (XAI). Two sources of data were used for the investigation: symptoms and feedback from vaccinated persons, and patient-specific data obtained from pathology-supported genetic testing (PSGT). The data was highly imbalanced and the two machine learning algorithms namely the Deep Neural Network(DNN) and the XGBoost classifier were trained with both the imbalanced data and the balanced data created by Synthetic Minority Oversampling Technique (SMOTE). Moreover the algorithms were trained with both the highly correlated features with breakthrough infection and all features. The DNN performed well when trained with all features and balanced data, it managed to score: f1-score 0.89, AUC 0.75, G-score 0.72 and MCC 0.58. XGBoost performed well when trained with highly correlated features and with both balanced and unbalanced data, it managed to score:  f1-score 0.88, AUC 0.75, G-score 0.72 and MCC 0.53. The kernelExplainer was used to identify the most important features for the DNN and TreeExplainer was used for the XGBoost model. The two explainers managed to identify that cholesterol level, body temperature, blood glucose level, physical activity per week, BP level, weight and waist circumference contributed much to breakthrough infection. These findings helps the medical practitioners to identify patient-specific variables that can be altered to prevent deterioration of patients.